// src/types/theme.ts
// material-ui
import {
  SimplePaletteColorOptions,
  PaletteColorOptions,
} from "@mui/material/styles"

// ==============================|| DEFAULT THEME - TYPES ||============================== //

export type PaletteThemeProps = {
  primary: SimplePaletteColorOptions
  secondary: SimplePaletteColorOptions
  error: SimplePaletteColorOptions
  warning: SimplePaletteColorOptions
  info: SimplePaletteColorOptions
  success: SimplePaletteColorOptions
  grey: PaletteColorOptions
}

export type CustomShadowProps = {
  button: string
  text: string
  z1: string
  primary: string
  primaryButton: string
  secondary: string
  secondaryButton: string
  error: string
  errorButton: string
  warning: string
  warningButton: string
  info: string
  infoButton: string
  success: string
  successButton: string
  grey: string
  greyButton: string
}